bpmn_pipeline/requirements.txt
docling
langchain
langchain-groq
langchain-core
pydantic>=2
lxml
networkx
python-dotenv
pytest
langfuse
langfuse